Major refactoring of existing and addition of new apps - scene manager works more smoothly   - light configs are now YAML instead of subclasses   - removed auto-reloading, as it did not work well   - introduced controller shortcuts via .controller_names.yaml - update to more modern version of HASS and AD - Added a simple Home Energy Management System (hems.py)

+++ apps/laundry_reminder.py
@@ -25,7 +25,7 @@ class LaundryReminder(hass.Hass):
 
     def set_wm_sensor(self):
 
-        power = int(self.get_state(self.power_sensor))
+        power = float(self.get_state(self.power_sensor))
 
         self.wm_state = "off"
         if power > 0:
